requirements.txt
fastapi
uvicorn
python-dotenv
pandas
langchain
langchain-community
langchain-google-genai
faiss-cpu
langchain-groq
